pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lumina"
version = "2.0.0"
description = "Lumina - Illuminate your memories. Organize and manage photo/video libraries with intelligent analysis and deduplication."
authors = [
    {name = "Ivan R. Judson", email = "irjudson@gmail.com"}
]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: System :: Archiving",
    "Topic :: Utilities",
]
dependencies = [
    "Pillow>=9.0.0,<12.0.0",
    "pillow-heif>=0.10.0,<1.0.0",
    "rawpy>=0.18.0,<1.0.0",
    "imageio>=2.25.0,<3.0.0",
    "pyexiftool>=0.5.0,<1.0.0",
    "click>=8.0.0,<9.0.0",
    "rich>=13.0.0,<14.0.0",
    "arrow>=1.2.0,<2.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "starlette>=0.40.0,<0.51.0",
    "httpx>=0.27.0,<0.28.0",
    "uvicorn[standard]>=0.24.0,<1.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "PyWavelets>=1.4.0,<2.0.0",
    "videohash>=3.0.0,<4.0.0",
    "sse-starlette>=1.6.0,<2.0.0",
    "watchdog>=3.0.0,<5.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "psycopg2-binary>=2.9.0,<3.0.0",
    "alembic>=1.12.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pgvector>=0.2.0,<1.0.0",
    "sqlmodel>=0.0.14,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<9.0.0",
    "pytest-cov>=4.0.0,<7.0.0",
    "pytest-benchmark>=4.0.0,<5.0.0",
    "pytest-xdist>=3.0.0,<4.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "black>=22.0.0,<25.0.0",
    "flake8>=4.0.0,<8.0.0",
    "flake8-docstrings>=1.7.0,<2.0.0",
    "flake8-bugbear>=24.0.0,<25.0.0",
    "mypy>=0.950,<2.0.0",
    "isort>=5.10.0,<6.0.0",
    "types-Pillow>=9.0.0,<11.0.0",
    "httpx>=0.24.0,<1.0.0",
    "pre-commit>=3.0.0,<5.0.0",
    "bandit[toml]>=1.7.0,<2.0.0",
]
gpu = [
    # GPU acceleration for perceptual hash computation
    # Install with: pip install -e ".[gpu]"
    # Or from PyPI index: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]
gpu-search = [
    # Fast similarity search with FAISS
    # Install with: pip install -e ".[gpu-search]"
    # Note: faiss-gpu not available via pip; using CPU version
    # For GPU support, install via conda: conda install -c conda-forge faiss-gpu
    "faiss-cpu>=1.7.4",
]
gpu-all = [
    # Complete GPU acceleration stack
    # Install with: pip install -e ".[gpu-all]"
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "faiss-cpu>=1.7.4",
]
tagging = [
    # AI-based auto-tagging with OpenCLIP
    # Install with: pip install -e ".[tagging]"
    "open-clip-torch>=2.24.0",
    "ftfy>=6.1.0",  # Text preprocessing for CLIP
]
ollama = [
    # Ollama client for LLaVA vision model integration
    # Install with: pip install -e ".[ollama]"
    # Requires Ollama server running locally: ollama serve
    "ollama>=0.3.0",
]
tagging-all = [
    # Complete tagging stack with both OpenCLIP and Ollama
    # Install with: pip install -e ".[tagging-all]"
    "open-clip-torch>=2.24.0",
    "ftfy>=6.1.0",
    "ollama>=0.3.0",
]

[project.scripts]
lumina-analyze = "lumina.cli.analyze:analyze"
lumina-organize = "lumina.cli.organize:organize"
lumina-web = "lumina.cli.web:web"
lumina-server = "lumina.cli.server:server"
lumina-generate-thumbnails = "lumina.cli.generate_thumbnails:generate"

[project.urls]
Homepage = "https://github.com/irjudson/lumina"
Repository = "https://github.com/irjudson/lumina"
Issues = "https://github.com/irjudson/lumina/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["lumina*"]
exclude = ["tests*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true

[tool.mypy]
python_version = "3.9"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_untyped_decorators = false
no_implicit_optional = false
warn_redundant_casts = false
warn_unused_ignores = false
warn_no_return = false
warn_unreachable = false
strict_equality = false
ignore_missing_imports = true
exclude = [
    'tests/',
]

[[tool.mypy.overrides]]
module = "exiftool.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pillow_heif.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pywt.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "torch.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "torchvision.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "faiss.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pyopencl.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "videohash.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "open_clip.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "ollama.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "ftfy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "rawpy.*"
ignore_missing_imports = true

# TODO: Fix these properly with correct type annotations
# These modules have complex SQLAlchemy/ORM patterns that need dedicated type fixes
[[tool.mypy.overrides]]
module = [
    "lumina.db.*",
    "lumina.core.catalog_config",
    "lumina.analysis.scanner",
    "lumina.analysis.duplicate_detector",
    "lumina.analysis.preview_extractor",
    "lumina.web.*",
    "lumina.api.routers.*",
    "lumina.cli.analyze",
    "lumina.cli.web",
    "lumina.jobs.tasks",
]
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = [
    "-n", "4",  # Parallel execution with pytest-xdist (4 workers for reasonable resource usage)
    "--tb=line",  # Line-only tracebacks for speed
    "-p", "no:warnings",  # Disable warnings to speed up output
    "--dist=loadfile",  # Distribute tests by file - ensures all tests in a file run on same worker
    "-m", "not e2e",  # Include integration tests but skip e2e (which require docker-compose up)
]
# Unique worker group name for isolation across multiple projects
xdist_worker_class = "lumina_workers"
markers = [
    "integration: marks tests as integration tests requiring database (PostgreSQL)",
    "e2e: marks tests as end-to-end tests requiring full Docker stack (docker-compose up)",
]
# Coverage disabled by default for speed - run with: pytest --cov=lumina
# To run with coverage: pytest --cov=lumina --cov-report=html
# Integration tests run by default (services always available on this machine)
# To skip integration tests: pytest -m "not integration"

[tool.coverage.run]
omit = [
    "tests/*",
]

[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv"]
skips = ["B101"]  # Skip assert_used in tests
